:art: Listing all possible events sent in Notification object
